+++ ru.taximaxim.codekeeper.core.tests/src/main/resources/ru/taximaxim/codekeeper/core/drop_ms_column_with_default_diff.sql
@@ -1,0 +1,7 @@
+ALTER TABLE [dbo].[Tab]
+	DROP CONSTRAINT [DF_Tab_DefData]
+GO
+
+ALTER TABLE [dbo].[Tab]
+	DROP COLUMN [DefData]
+GO