+++ CHANGELOG.md
@@ -9,6 +9,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ### Added
 
+- Added localization of information messages in the CLI version.
+
 ### Changed
 
 ### Fixed