+++ CHANGELOG.md
@@ -9,6 +9,16 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ### Added
 
+### Changed
+
+### Fixed
+
+### Removed
+
+## [10.3.0] - 2024-11-12
+
+### Added
+
 - Added localization of information messages in the CLI version.
 
 ### Changed
@@ -19,8 +29,6 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 - Fixed code generation when deleting column with a default value in MS SQL.
 - Fixed a bug when reading from a database in MS SQL.
-
-### Removed
 
 ## [10.2.0] - 2024-10-29
 
@@ -2098,7 +2106,8 @@ We increment major version in this release due to decision to shift our developm
 
 - ORDER BY parsing has been added to the aggregate functions.
 
-[Unreleased]: https://github.com/pgcodekeeper/pgcodekeeper/compare/v10.2.0...HEAD
+[Unreleased]: https://github.com/pgcodekeeper/pgcodekeeper/compare/v10.3.0...HEAD
+[10.3.0]: https://github.com/pgcodekeeper/pgcodekeeper/compare/v10.2.0...v10.3.0
 [10.2.0]: https://github.com/pgcodekeeper/pgcodekeeper/compare/v10.1.0...v10.2.0
 [10.1.0]: https://github.com/pgcodekeeper/pgcodekeeper/compare/v10.0.0...v10.1.0
 [10.0.0]: https://github.com/pgcodekeeper/pgcodekeeper/compare/v9.14.0...v10.0.0